+++ opentelemetry/src/metrics/mod.rs
@@ -6,7 +6,9 @@ mod instruments;
 mod meter;
 pub mod noop;
 #[cfg(feature = "experimental_metrics_bound_instruments")]
-pub use instruments::{counter::BoundCounter, histogram::BoundHistogram, BoundSyncInstrument};
+pub use instruments::{
+    counter::BoundCounter, gauge::BoundGauge, histogram::BoundHistogram, BoundSyncInstrument,
+};
 pub use instruments::{
     counter::{Counter, ObservableCounter},
     gauge::{Gauge, ObservableGauge},